tests/requirements.txt
# Test dependencies for RDF Converter
# Install with: pip install -r tests/requirements.txt

# Core testing framework
pytest>=7.0.0
pytest-logging>=2015.11.4

# Required for the application (if not already installed)
rdflib>=6.0.0
pyyaml>=6.0
requests>=2.28.0

# Optional but recommended for better test output
pytest-cov>=3.0.0
pytest-html>=3.1.0
pytest-xdist>=2.5.0  # For parallel test execution
